avoid extra newlines in diff (working around upstream wdiff bug)
Multiline substitutions are surprisingly clumsy, with or without perl,
so I trusted the least incomprehensible (if hackish) recipe from
stackoverflow; it's a hack which assumes no '\r' in the input file, and
checking that is *also* surprisingly clumsy, so I resorted to checking
that the file stays the same after deleting '\r'.

Any cleaner solution is welcome.

Closes #9